+++ documentation-site/src/pages/demos/moving-platform/_create-boundaries.ts
@@ -49,6 +49,7 @@ export async function createBoundaries(
     getAssetUrl('img/White.png'),
   );
   const wallSprite = createImageSprite(wallImage, renderContext, {
+    pixelsPerUnit: 1,
     layer: renderLayer,
   });